requirements.txt
langchain==0.2.1
langchain_google_genai==1.0.4
python-dotenv==1.0.1